Move MovePDFTemplatesView to templates.py
